requirements.txt
flask
pymongo
python-dotenv
flask-wtf
werkzeug
Flask-Login
email-validator
flask-mail